started heartbeat logic

+++ Assets/Version1/Nats/Messages/Host/CreateSessionMessage.cs
@@ -3,7 +3,7 @@ namespace Version1.Nats.Messages.Host
     public class CreateSessionMessage : BaseMessage
     {
         public int SessionToken;
-        public CreateSessionMessage(string dateTime, int lobbyID, byte playerID, int sessionToken) : base(dateTime, MessageSubject.CreateSession, lobbyID, playerID)
+        public CreateSessionMessage(string dateTime, int lobbyID, int playerID, int sessionToken) : base(dateTime, MessageSubject.CreateSession, lobbyID, playerID)
         {
             SessionToken = sessionToken;
         }
